+++ src/components/symptomTypePage/SymptomTypeComponent.tsx
