+++ src/pages/DiagnosisList.tsx
@@ -5,10 +5,11 @@ import MainHeader from "../components/header/MainHeader";
 import ListComponent from "../components/diagnosisListPage/ListComponent";
 import theme from "../lib/theme";
 import { useNavigate } from "react-router-dom";
-import { Body_2, Description, Heading_3 } from "../lib/fontStyle";
+import { Description, Heading_3 } from "../lib/fontStyle";
 import { useEffect } from "react";
 import axios from "axios";
 import { IDiagnosisList } from "../interfaces/component";
+import EmptyPage from "../components/diagnosisListPage/EmptyPage";
 
 const Container = styled.section`
   padding-top: 9.6rem;
@@ -49,13 +50,6 @@ const Highlight = styled.span<{ type: string }>`
 const List = styled.section`
   margin: 0 2.4rem 10rem 2.4rem;
 `;
-const EmptyContainer = styled.section`
-  padding-top: 12.5rem;
-`;
-const EmptyText = styled(Body_2)`
-  text-align: center;
-  color: ${({ theme }) => theme.color.grey_400};
-`;
 
 const DiagnosisList = () => {
   const navigate = useNavigate();
@@ -89,17 +83,7 @@ const DiagnosisList = () => {
     <Container>
       <MainHeader />
       {diagnosisList.length === 0 ? (
-        <>
-          <Title>
-            빠른 진단으로
-            <br /> 내 몸의 정확한 증상을
-            <br />
-            알아보세요!
-          </Title>
-          <EmptyContainer>
-            <EmptyText>진단 내역이 없어요</EmptyText>
-          </EmptyContainer>
-        </>
+        <EmptyPage />
       ) : (
         <>
           <Title>
