+++ src/pages/SymptomTypePage.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import styled from "styled-components";
 import ContentHeader from "../components/header/ContentHeader";
-import SymptomTypeComponent from "../components/selectSymptomPage/SymptomTypeComponent";
+import SymptomTypeComponent from "../components/symptomTypePage/SymptomTypeComponent";
 import { Heading_3 } from "../lib/fontStyle";
 import { symptom_type } from "../store/symptom_type";
 import SymptomModal from "../components/modal/SymptomModal";
@@ -36,7 +36,7 @@ const SymptomContainer = styled.section`
   }
 `;
 
-const SelectSymptomPage = () => {
+const SymptomTypePage = () => {
   const [select, setSelect] = useState(-1);
   const [modal, setModal] = useState(false);
 
@@ -70,4 +70,4 @@ const SelectSymptomPage = () => {
   );
 };
 
-export default SelectSymptomPage;
+export default SymptomTypePage;
